requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
defusedxml==0.6.0
Django==3.0.8
django-allauth==0.42.0
django-extra-fields==2.0.5
django-hitcount==1.3.2
djangorestframework==3.11.0
httpie==2.2.0
idna==2.10
oauthlib==3.1.0
Pillow==7.2.0
Pygments==2.6.1
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
sqlparse==0.3.1
urllib3==1.25.9
